requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
decorator==5.2.1
fonttools==4.61.1
future==1.0.0
geocoder==1.38.1
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.13.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.2
pyarrow==22.0.0
pydeck==0.9.1
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
pytz==2025.2
ratelim==0.1.6
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
smmap==5.0.2
streamlit==1.52.1
tenacity==9.1.2
tkintermapview==1.29
toml==0.10.2
tornado==6.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
